Imrove migrate index file field. change name use reference and add an option --missing-content-only
